AtomicPosition: fix overflowing addition

+++ src/draw_target.rs
@@ -351,7 +351,7 @@ impl RateLimiter {
 
         // We add `new` to `capacity`, subtract one for returning `true` from here,
         // then make sure it does not exceed a maximum of `MAX_BURST`, then store it.
-        self.capacity = Ord::min(MAX_BURST, self.capacity + new as u8 - 1);
+        self.capacity = Ord::min(MAX_BURST as u128, (self.capacity as u128) + new - 1) as u8;
         // Store `prev` for the next iteration after subtracting the `remainder`.
         self.prev = now - Duration::from_nanos(remainder as u64);
         true
